// src/types/navigation.ts
export type RootStackParamList = {
    Home: undefined;
    AlgorithmDetail: {
      id: string;
      title: string;
      category: string;
    };
    Settings: undefined;
  };
  
  export type TabParamList = {
    Discover: undefined;
    Favorites: undefined;
    Profile: undefined;
  };